refactor: remove duplicate Document model
