--- src/commonMain/kotlin/com/xebia/functional/domain/Document.kt
@@ -1,3 +1,0 @@
-package com.xebia.functional.domain
-
-data class Document(val content: String)
